actions: Split out zerver.lib.user_counts.
(cherry picked from commit 703186c339a10722415a7d78a443385b7ed9c833)
