+++ zerver/lib/actions.py
@@ -29,7 +29,7 @@ from django.conf import settings
 from django.contrib.contenttypes.models import ContentType
 from django.core.exceptions import ValidationError
 from django.db import IntegrityError, connection, transaction
-from django.db.models import Count, Exists, F, OuterRef, Q, Sum
+from django.db.models import Exists, F, OuterRef, Q, Sum
 from django.db.models.query import QuerySet
 from django.utils.html import escape
 from django.utils.timezone import now as timezone_now
@@ -193,6 +193,7 @@ from zerver.lib.upload import (
     delete_message_image,
     upload_emoji_image,
 )
+from zerver.lib.user_counts import realm_user_count, realm_user_count_by_role
 from zerver.lib.user_groups import (
     access_user_group_by_id,
     create_system_user_groups_for_realm,
@@ -319,31 +320,6 @@ def create_historical_user_messages(*, user_id: int, message_ids: List[int]) -> 
 
 def subscriber_info(user_id: int) -> Dict[str, Any]:
     return {"id": user_id, "flags": ["read"]}
-
-
-def realm_user_count(realm: Realm) -> int:
-    return UserProfile.objects.filter(realm=realm, is_active=True, is_bot=False).count()
-
-
-def realm_user_count_by_role(realm: Realm) -> Dict[str, Any]:
-    human_counts = {
-        str(UserProfile.ROLE_REALM_ADMINISTRATOR): 0,
-        str(UserProfile.ROLE_REALM_OWNER): 0,
-        str(UserProfile.ROLE_MODERATOR): 0,
-        str(UserProfile.ROLE_MEMBER): 0,
-        str(UserProfile.ROLE_GUEST): 0,
-    }
-    for value_dict in list(
-        UserProfile.objects.filter(realm=realm, is_bot=False, is_active=True)
-        .values("role")
-        .annotate(Count("role"))
-    ):
-        human_counts[str(value_dict["role"])] = value_dict["role__count"]
-    bot_count = UserProfile.objects.filter(realm=realm, is_bot=True, is_active=True).count()
-    return {
-        RealmAuditLog.ROLE_COUNT_HUMANS: human_counts,
-        RealmAuditLog.ROLE_COUNT_BOTS: bot_count,
-    }
 
 
 def get_signups_stream(realm: Realm) -> Stream:
